% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021.bkc_bkb.R
\name{bkc_bkb}
\alias{bkc_bkb}
\title{Background correction for the backbone markers}
\usage{
bkc_bkb(
  paths,
  bkb.v,
  bkb.upper.quntile = 0.9,
  bkb.lower.quntile = 0.1,
  bkb.min.quntile = 0.01
)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{bkb.upper.quntile}{The cut-off (default = 0.9) for selecting cells used for estimating the parameter of signal.}

\item{bkb.lower.quntile}{The cut-off (default = 0.1) for selecting cells used for estimating the parameters of noise.}

\item{bkb.min.quntile}{The cut-off (default = 0.01) for omitting the cells with the smallest values to minimise the impact of outliers.}
}
\value{
Generating the calibrated measurements and save to medpara_bkc.bkb_no.bkcPhy_mt.rds file. Visualising the result with the scatter plots.
}
\description{
This function has been designed to do background correction for the backbone markers by using normal-exponential convolution model.
}
\author{
Hsiao-Chi Liao and Agus Salim
}
