% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022.bkc_pe.R
\name{bkc_pe}
\alias{bkc_pe}
\title{Background correction for the well-specific markers (PE)}
\usage{
bkc_pe(paths, pe.mean.sd = 3, pe.lower.quntile = 0.1, pe.min.quntile = 0.01)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{pe.mean.sd}{Selecting cells with the value larger than mean + ?sd (default = 3) to estimate the parameter of signal.}

\item{pe.lower.quntile}{The cut-off (default = 0.1) for selecting cells used for estimating the parameters of noise.}

\item{pe.min.quntile}{The cut-off (default = 0.01) for omitting the cells with the smallest values to minimise the impact of outliers.}
}
\value{
Generating the calibrated measurements and save to bkc.pe_mt.rds file. Visualising the result with the scatter plots.
}
\description{
This function has been designed to do background correction for the well-specific markers (PE) by using normal-exponential convolution model.
}
\author{
Hsiao-Chi Liao and Agus Salim
}
