% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070.cluster.analysis_bkbOnly.R
\name{cluster.analysis.bkbOnly}
\alias{cluster.analysis.bkbOnly}
\title{Cluster analysis with normalised backbone measurements and the complete dataset}
\usage{
cluster.analysis.bkbOnly(paths, bkb.v)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}
}
\value{
Updating the metadata for cells in the fcs_metadata_df.rds file, adding the information of the biological clusters from the clean and complete dataset. Visualising the result with the scatter plots.
}
\description{
This function has been designed to perform cluster analysis for the normalised backbone measurements and the complete dataset which includes the normalised backbone measurements and the imputed well-specific markers.
}
\author{
Hsiao-Chi Liao
}
