% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060.cluster.analysis.R
\name{cluster.analysis}
\alias{cluster.analysis}
\title{Cluster analysis with normalised backbone measurements and the complete dataset}
\usage{
cluster.analysis(paths, bkb.v, yvar = "Legend", control.wells)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{yvar}{The name of the well-specific marker in the FCS files (e.g., "Legend").}

\item{control.wells}{The well label of the control wells, including the autofluorescence and the isotype controls (format: plate_well, e.g., P1_A01)}
}
\value{
Updating the metadata for cells in the fcs_metadata_df.rds file, adding the information of the biological clusters from the clean and complete dataset. Visualising the result with the scatter plots.
}
\description{
This function has been designed to perform cluster analysis for the normalised backbone measurements and the complete dataset which includes the normalised backbone measurements and the imputed well-specific markers.
}
\author{
Hsiao-Chi Liao
}
