% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011.fcs_to_rds.R
\name{fcs_to_rds}
\alias{fcs_to_rds}
\title{Converting FCS files to RDS files}
\usage{
fcs_to_rds(paths, file_meta, yvar)
}
\arguments{
\item{paths}{a vector of characters of paths to store intput, intermediary results, outputs...}

\item{file_meta}{if the file names of the FCS files are in the specified format, set file_meta="auto"; otherwise set file_meta="usr" and provide "filename_meta.csv" in FCSpath.}

\item{yvar}{the name of the well-specific marker in the FCS files (e.g., "Legend").}
}
\value{
Raw protein intensities and the corresponding metadata from MPC experiments
}
\description{
This function has been designed to convert the raw FCS files to data matrix and export to RDS files.
}
\details{
Generating fcs_metadata_df.rds and fcs_rawInten_mt.rds files in the output directory.
}
\author{
Hsiao-Chi Liao
}
