% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012.fcs_to_rds_bkb.R
\name{fcs_to_rds_bkb}
\alias{fcs_to_rds_bkb}
\title{Converting FCS files to RDS files (for the case without exploratory markers)}
\usage{
fcs_to_rds_bkb(paths, file_meta = "auto", plate_based = TRUE)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{file_meta}{If the file names of the FCS files are in the specified format, set file_meta="auto"; otherwise set file_meta="usr" and provide "filename_meta.csv" in FCSpath.}

\item{plate_based}{If the data is from the plate-based assay, set plate_based = TRUE. Setting FALSE represents data from the fluorescence flow cytometry (FFC) assay.}
}
\value{
Generating fcs_metadata_df.rds and fcs_rawInten_mt.rds files.
}
\description{
This function has been designed to convert the raw FCS files to data matrix and export to RDS files.
}
\author{
Hsiao-Chi Liao
}
