% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052.imputation_mice.R
\name{imputation_mice}
\alias{imputation_mice}
\title{Imputing the unmeasured well-specific markers with MICE framework}
\usage{
imputation_mice(
  paths,
  chans,
  yvar = "Legend",
  cores = cores,
  iter = 5,
  models.use = models.use,
  extra_args_regression_params = extra_args_regression_params,
  control.wells = control.wells,
  prediction_events_downsampling = NULL
)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{chans}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{yvar}{The name of the well-specific marker in the FCS files (has been changed to "Legend" in the first function).}

\item{cores}{The number of cores used to perform parallel computation (default = 8L).}

\item{iter}{The number of iterations for MICE to update the imputations}

\item{models.use}{A vector of the names of the models used for imputation (an example: c("LM", "LASSO3", "SVM", "XGBoost")). The length of the vector is arbitrary.}

\item{extra_args_regression_params}{A list of the lists of the parameters for running regression models.}

\item{control.wells}{The well label of the control wells, including the autofluorescence and the isotype controls (format: plate_well, e.g., P1_A01)}

\item{prediction_events_downsampling}{Default = NULL (not doing subsampling). How many cells per well you want to have the imputation? (must be less than or equal to a half as we won't get the prediction for cells in the training set).}
}
\value{
Imputing the unmeasured well-specific markers and save to predictions.Rds file. Visualising the result with the boxplots (r-sq).
}
\description{
This function has been designed to impute/predict the unmeasured well-specific markers with MICE framework.
}
\author{
Hsiao-Chi Liao and InfinityFlow (Becht et. al, 2021)
}
