% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031.initM.R
\name{initM}
\alias{initM}
\title{Initial biological clusters}
\usage{
initM(paths, bkb.v, trans.dat = "cent.lgc")
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{trans.dat}{Method for transforming raw intensities (default: Logicle).}
}
\value{
Updating the metadata for cells in the fcs_metadata_df.rds file, adding the information of the initial biological clusters. Visualising the result with the scatter plots.
}
\description{
Generating the M matrix for removing well effects.
}
\details{
This function has been designed to find initial biological clusters with centred transformed data.
}
\author{
Hsiao-Chi Liao
}
