% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuiltinDataDoc.R
\docType{data}
\name{ord.fcs.raw.meta.df.out_ffc}
\alias{ord.fcs.raw.meta.df.out_ffc}
\title{Metadata of the example MPC data}
\format{
a data.frame
}
\usage{
data(ord.fcs.raw.meta.df.out_ffc)
}
\description{
Metadata of the example MPC data
}
\keyword{datasets}
