% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuiltinDataDoc.R
\docType{data}
\name{ord.fcs.raw.meta.df.out_mpc}
\alias{ord.fcs.raw.meta.df.out_mpc}
\title{Metadata of the example FFC data}
\format{
a data.frame
}
\usage{
data(ord.fcs.raw.meta.df.out_mpc)
}
\description{
Metadata of the example FFC data
}
\keyword{datasets}
