% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuiltinDataDoc.R
\docType{data}
\name{ord.fcs.raw.mt_ffc}
\alias{ord.fcs.raw.mt_ffc}
\title{Subset of the single-cell murine lung data at steady state from an MPC experiment}
\format{
a matrix containing cells from 266 wells (50 cells/well)
}
\source{
\url{https://flowrepository.org/id/FR-FCM-Z2LP}
}
\usage{
data(ord.fcs.raw.mt_ffc)
}
\description{
Subset of the single-cell murine lung data at steady state from an MPC experiment
}
\keyword{datasets}
