% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuiltinDataDoc.R
\docType{data}
\name{ord.fcs.raw.mt_mpc}
\alias{ord.fcs.raw.mt_mpc}
\title{Subset of the sorted CD4+ and CD8+ T cells from mice splenocytes from an FFC experiment}
\format{
a matrix containing cells from 5 batches (50 cells/batch)
}
\source{
\url{http://flowrepository.org/id/FR-FCM-Z6UG}
}
\usage{
data(ord.fcs.raw.mt_mpc)
}
\description{
Subset of the sorted CD4+ and CD8+ T cells from mice splenocytes from an FFC experiment
}
\keyword{datasets}
