% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/032.rmWellEffect.R
\name{rmWellEffect}
\alias{rmWellEffect}
\title{Removing well effect from the data}
\usage{
rmWellEffect(paths, visualisation = TRUE)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{visualisation}{Whether to visualise the results by using heatmap (default=TRUE).}
}
\value{
Generating the calibrated measurements and save to bkc.adj.bkb_logScale_mt.rds (on log scale) and bkc.adj.bkb_linearScale_mt.rds (on linear scale). Visualising the result with the heatmaps.
}
\description{
This function has been designed to remove the unwanted effects (well effects) from the background corrected measurements.
}
\author{
Hsiao-Chi Liao
}
