% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000.runMAPFX.R
\name{runMAPFX}
\alias{runMAPFX}
\title{MAssively Parallel Flow cytometry Xplorer (MAPFX) for analysing single-cell protein intensity from the Massively-Parallel Cytometry (MPC) Experiments - imputation using backbone markers}
\usage{
runMAPFX(FCSpath="/PathToFCSfiles/",
Outpath="/PathToOutputFolder/Output/",
file_meta="auto",
bkb.v = c("FSC-H", "FSC-W", "SSC-H", "SSC-W", "CD69-CD301b", "MHCII", "CD4", "CD44",
          "CD8", "CD11c", "CD11b", "F480", "Ly6C", "Lineage", "CD45a488",
          "CD24", "CD103"),
chans = c("FSC-H", "FSC-W", "SSC-H", "SSC-W", "CD69-CD301b", "MHCII", "CD4", "CD44",
          "CD8", "CD11c", "CD11b", "F480", "Ly6C", "Lineage", "CD45a488",
          "CD24", "CD103"),
yvar="Legend", control.wells = c("P1_A01", "P2_A02", "P3_G02"),
bkb.upper.quntile=0.9, bkb.lower.quntile=0.1, bkb.min.quntile=0.01,
pe.mean.sd=3, pe.lower.quntile=0.1, pe.min.quntile=0.01,
trans.dat="cent.lgc",
models.use = c("XGBoost"),
extra_args_regression_params = list(list(nrounds = 1500, eta = 0.03)),
cores=2L)
}
\arguments{
\item{FCSpath}{Path to the input directory where input FCS files are stored (one file per well). Will look for FCS files recursively in that directory.}

\item{Outpath}{Path to the output directory where intermediate results and final results will be stored}

\item{file_meta}{If the file names of the FCS files are in the specified format, set file_meta="auto"; otherwise set file_meta="usr" and provide "filename_meta.csv" in FCSpath.}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{chans}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{yvar}{The name of the well-specific marker in the FCS files (e.g., "Legend").}

\item{control.wells}{The well label of the control wells, including the autofluorescence and the isotype controls (format: plate_well, e.g., P1_A01)}

\item{bkb.upper.quntile}{The cut-off (default = 0.9) for selecting cells used for estimating the parameter of signal.}

\item{bkb.lower.quntile}{The cut-off (default = 0.1) for selecting cells used for estimating the parameters of noise.}

\item{bkb.min.quntile}{The cut-off (default = 0.01) for omitting the cells with the smallest values to minimise the impact of outliers.}

\item{pe.mean.sd}{Selecting cells with the value larger than mean + ?sd (default = 3) to estimate the parameter of signal.}

\item{pe.lower.quntile}{The cut-off (default = 0.1) for selecting cells used for estimating the parameters of noise.}

\item{pe.min.quntile}{The cut-off (default = 0.01) for omitting the cells with the smallest values to minimise the impact of outliers.}

\item{trans.dat}{Method for transforming raw intensities (default: Logicle).}

\item{models.use}{A vector of the names of the models used for imputation (an example: c("LM", "LASSO3", "SVM", "XGBoost")). The length of the vector is arbitrary.}

\item{extra_args_regression_params}{A list of the lists of the parameters for running regression models. (default = list(list(degree = 1), list(nfolds = 10, degree = 3),
list(type = "nu-regression", cost = 8, nu = 0.5, kernel = "radial"),
list(nrounds = 1500, eta = 0.03)))}

\item{cores}{The number of cores used to perform parallel computation (default = 4L).}
}
\value{
Normalised backbone measurements and imputed well-specific markers. Cluster analysis for both normalised backbones and the completed dataset. Graphs will be provided.
}
\description{
This function is an end-to-end toolbox for analysing the data from the LEGENDScreen assay in FCS format, including data normalisation, imputation and cluster analysis.
}
\author{
Hsiao-Chi Liao, Agus Salim, and InfinityFlow (Becht et. al, 2021)
}
