% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000.runMAPFX_bkbOnly.R
\name{runMAPFX_bkbOnly}
\alias{runMAPFX_bkbOnly}
\title{MAssively Parallel Flow cytometry Xplorer (MAPFX) for analysing single-cell protein intensity from fluorescence flow cytometry (FFC).}
\usage{
runMAPFX_bkbOnly(FCSpath="/PathToFCSfiles/",
Outpath="/PathToOutputFolder/Output/",
file_meta="auto",
plate_based = TRUE,
bkb.v = c("FSC-H", "FSC-W", "SSC-H", "SSC-W", "CD69-CD301b", "MHCII", "CD4", "CD44",
          "CD8", "CD11c", "CD11b", "F480", "Ly6C", "Lineage", "CD45a488",
          "CD24", "CD103"),
bkb.upper.quntile=0.9, bkb.lower.quntile=0.1, bkb.min.quntile=0.01,
trans.dat="cent.lgc")
}
\arguments{
\item{FCSpath}{Path to the input directory where input FCS files are stored (one file per well). Will look for FCS files recursively in that directory.}

\item{Outpath}{Path to the output directory where intermediate results and final results will be stored}

\item{file_meta}{If the file names of the FCS files are in the specified format, set file_meta="auto"; otherwise set file_meta="usr" and provide "filename_meta.csv" in FCSpath.}

\item{plate_based}{If the data is from the plate-based assay, set plate_based = TRUE. Setting FALSE represents data from the fluorescence flow cytometry (FFC) assay.}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{bkb.upper.quntile}{The cut-off (default = 0.9) for selecting cells used for estimating the parameter of signal.}

\item{bkb.lower.quntile}{The cut-off (default = 0.1) for selecting cells used for estimating the parameters of noise.}

\item{bkb.min.quntile}{The cut-off (default = 0.01) for omitting the cells with the smallest values to minimise the impact of outliers.}

\item{trans.dat}{Method for transforming raw intensities (default: Logicle).}
}
\value{
Normalised backbone measurements. Cluster analysis for normalised backbones. Graphs will be provided.
}
\description{
This function is an end-to-end toolbox for analysing the data from FFC in FCS format, including data normalisation and cluster analysis.
}
\author{
Hsiao-Chi Liao, Agus Salim
}
