% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070.cluster.analysis_bkbOnly.R
\name{cluster.analysis.bkbOnly}
\alias{cluster.analysis.bkbOnly}
\title{Cluster analysis with normalised backbone measurements}
\usage{
cluster.analysis.bkbOnly(paths, bkb.v, plots = TRUE)
}
\arguments{
\item{paths}{a vector of characters of paths to store intput, intermediary results, outputs...}

\item{bkb.v}{a vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{plots}{logical; if TRUE (default), produce an UMAP embedding plot from the normalised backbone markers to visualise the structure of the biological clusters.}
}
\value{
Updating the metadata for cells in the fcs_metadata_df.rds file, adding the information of the biological clusters from the clean and complete dataset. Visualising the result with the scatter plots.
}
\description{
This function has been designed to perform cluster analysis for the normalised backbone measurements.
}
\author{
Hsiao-Chi Liao
}
