% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031.initM.R
\name{initM}
\alias{initM}
\title{Initial biological clusters}
\usage{
initM(paths, assay, bkb.v, plots = TRUE)
}
\arguments{
\item{paths}{a vector of characters of paths to store intput, intermediary results, outputs...}

\item{assay}{the type of the input data - MPC or FFC.}

\item{bkb.v}{a vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{plots}{logical; if TRUE (default), produce a UMAP embedding plot to visualise the structure of the biological clusters to form the initial M matrix for removal of well effect.}
}
\value{
Metadata for cells with the initial biological clusters labels added
}
\description{
Generating the M matrix for removing well effects.
}
\details{
This function has been designed to find initial biological clusters with centred transformed data.

Updating the metadata for cells in the fcs_metadata_df.rds file, adding the information of the initial biological clusters, and visualising the result with the scatter plots in the output directory.
}
\author{
Hsiao-Chi Liao
}
