% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/033.rmBatchEffect.R
\name{rmBatchEffect}
\alias{rmBatchEffect}
\title{Removing batch effect from the data (FFC)}
\usage{
rmBatchEffect(paths, plots = TRUE)
}
\arguments{
\item{paths}{a vector of characters of paths to store intput, intermediary results, outputs...}

\item{plots}{logical; if TRUE (default), produce heatmaps to visualise the unwanted (batch) effects and biological effects in the pre- and post- adjusted datasets.}
}
\value{
Generating the calibrated measurements and save to bkc.adj.bkb_logScale_mt.rds (on log scale) and bkc.adj.bkb_linearScale_mt.rds (on linear scale). Visualising the result with the heatmaps.
}
\description{
This function has been designed to remove the unwanted effects (batch effects) from the background corrected measurements.
}
\author{
Hsiao-Chi Liao
}
